+++ src/main/java/org/mule/module/hubspot/client/HubSpotClient.java
@@ -9,6 +9,7 @@
 
 package org.mule.module.hubspot.client;
 
+import java.util.List;
 import java.util.Map;
 
 import org.mule.module.hubspot.exception.HubSpotConnectorAccessTokenExpiredException;
@@ -21,6 +22,7 @@ import org.mule.module.hubspot.model.contact.ContactList;
 import org.mule.module.hubspot.model.contact.ContactProperties;
 import org.mule.module.hubspot.model.contact.ContactQuery;
 import org.mule.module.hubspot.model.contact.ContactStatistics;
+import org.mule.module.hubspot.model.contactproperty.CustomContactProperty;
 import org.mule.module.hubspot.model.list.HubSpotList;
 import org.mule.module.hubspot.model.list.HubSpotListLists;
 
@@ -71,5 +73,8 @@ public interface HubSpotClient {
 			throws HubSpotConnectorException, HubSpotConnectorNoAccessTokenException, HubSpotConnectorAccessTokenExpiredException;
 	
 	public ContactList getContactsInAList(String accessToken, String userId, String listId, String count, String property, String offset)
+			throws HubSpotConnectorException, HubSpotConnectorNoAccessTokenException, HubSpotConnectorAccessTokenExpiredException;
+	
+	public List<CustomContactProperty> getAllProperties(String accessToken, String userId) 
 			throws HubSpotConnectorException, HubSpotConnectorNoAccessTokenException, HubSpotConnectorAccessTokenExpiredException;
 }