+++ src/test/java/org/mule/module/hubspot/unit/ContactsApiJacksonParsingTest.java
@@ -28,6 +28,7 @@ import org.mule.module.hubspot.model.contact.ContactList;
 import org.mule.module.hubspot.model.contact.ContactListMembership;
 import org.mule.module.hubspot.model.contact.ContactProperties;
 import org.mule.module.hubspot.model.contact.ContactStatistics;
+import org.mule.module.hubspot.model.contactproperty.CustomContactProperty;
 import org.mule.module.hubspot.model.list.HubSpotList;
 import org.mule.module.hubspot.model.list.HubSpotListFilter;
 import org.mule.module.hubspot.model.list.HubSpotListLists;
@@ -113,6 +114,19 @@ public class ContactsApiJacksonParsingTest {
 	}
 	
 	@Test
+	public void getContactByIdOutput2() throws JsonParseException, JsonMappingException, IOException {
+		InputStream input = Thread.currentThread().getContextClassLoader().getResourceAsStream(getContatPathFor("getContactByIdOutput2.json"));
+
+		Contact c = objMapper.readValue(input, Contact.class); 
+		ContactProperties cp = c.getProperties();
+		
+		Assert.assertNotNull(c);
+		Assert.assertFalse(StringUtils.isEmpty(cp.getFirstname()));
+		Assert.assertFalse(StringUtils.isEmpty(cp.getLastname()));
+		Assert.assertFalse(StringUtils.isEmpty(cp.getEmail()));
+	}
+	
+	@Test
 	public void getContactStatisticsOutput() throws JsonParseException, JsonMappingException, IOException {
 		InputStream input = Thread.currentThread().getContextClassLoader().getResourceAsStream(getContatPathFor("getContactstatisticsOutput.json"));
 		
@@ -168,11 +182,28 @@ public class ContactsApiJacksonParsingTest {
 		Assert.assertEquals(hblf.getList(), "1");
 	}
 	
+	@Test
+	public void getAllContactPropertiesFromAPortalOutput() throws JsonParseException, JsonMappingException, IOException {
+		InputStream input = Thread.currentThread().getContextClassLoader().getResourceAsStream(getContactPropertyPathFor("getAllContactPropertiesFromAPortalOutput.json"));
+		
+		CustomContactProperty[] lcp = objMapper.readValue(input, CustomContactProperty[].class);
+		CustomContactProperty cp = lcp[0];
+		
+		Assert.assertNotNull(lcp);
+		Assert.assertTrue(lcp.length > 0);
+		Assert.assertNotNull(cp);
+		Assert.assertEquals(cp.getName(), "billing_address1");
+	}
+	
 	private String getContatPathFor(String jsonFile) {
 		return String.format("contacts%s%s", File.separator, jsonFile);
 	}
 	
 	private String getListPathFor(String jsonFile) {
 		return String.format("lists%s%s", File.separator, jsonFile);
+	}
+	
+	private String getContactPropertyPathFor(String jsonFile) {
+		return String.format("contactproperty%s%s", File.separator, jsonFile);
 	}
 }